Create cognito stack to create a user pool which will be shared between other stacks
Client ID and Userpool ID will be used in the creation of the auth app stack and the app stack sourced from a config file. Config file added to gitignore. Required updating tsconfig.json to set esModuleInterop as true. Both API's testing and working as expected using the one user pool.
Cognito stack set to do not delete as is userpool